+++ app/(auth)/layout.tsx
@@ -1,0 +1,14 @@
+import Background from '@/components/shared/background'
+
+export default async function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <main className='flex h-dvh w-full flex-col items-center justify-center bg-center p-4 max-sm:pb-8'>
+      <Background />
+      {children}
+    </main>
+  )
+}
