+++ app/(auth)/sign-up/[[...sign-up]]/page.tsx
@@ -1,11 +1,9 @@
 import Loading from '@/app/loading'
-import Background from '@/components/shared/background'
 import { ClerkLoaded, ClerkLoading, SignUp } from '@clerk/nextjs'
 
 export default function Page() {
   return (
-    <main className='flex h-dvh w-full flex-col items-center justify-center bg-center p-4'>
-      <Background />
+    <>
       <ClerkLoading>
         <Loading />
       </ClerkLoading>
@@ -14,6 +12,6 @@ export default function Page() {
           <SignUp />
         </div>
       </ClerkLoaded>
-    </main>
+    </>
   )
 }
