+++ app/page.tsx
@@ -7,9 +7,6 @@ export default async function Home() {
     <main className='flex h-dvh w-full flex-col items-center justify-center bg-center'>
       <Background />
       <WelcomeScreen />
-      <footer className='mt-4 hidden sm:block'>
-        <FooterNarrow />
-      </footer>
     </main>
   )
 }
