+++ components/admin/adminAlert.tsx
@@ -1,5 +1,5 @@
 import { IconAlertCircle } from '@tabler/icons-react'
-import { Alert, AlertDescription, AlertTitle } from '../ui/alert'
+import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
 
 export default function AdminAlert({
   title,
