+++ components/admin/adminUserColumns.tsx
@@ -4,7 +4,7 @@ import { Button } from '@/components/ui/button'
 import { DataTableColumnHeader } from '@/components/ui/data-table-column-header'
 import { SelectUser } from '@/db/schema'
 import { ColumnDef } from '@tanstack/react-table'
-import { Badge } from '../ui/badge'
+import { Badge } from '@/components/ui/badge'
 import { cn } from '@/lib/utils'
 import { calculateAge } from '@/lib/birthday'
 import { IconCopy, IconTrash } from '@tabler/icons-react'
