+++ components/admin/manualAttendanceTab.tsx
@@ -1,4 +1,4 @@
-import { TabsContent } from '../ui/tabs'
+import { TabsContent } from '@/components/ui/tabs'
 
 export default function ManualAttendanceTab() {
   return (
