+++ components/admin/qrScannerTab.tsx
@@ -13,7 +13,7 @@ import { toast } from 'sonner'
 import { useUser } from '@clerk/nextjs'
 import { registerAttendanceRecord } from '@/actions/attendance'
 import { FetchAttendanceProps } from '@/types'
-import { ScanErrorScreen, ScanSuccessScreen } from './scanStateScreen'
+import { ScanErrorScreen, ScanSuccessScreen } from '@/components/admin/scanStateScreen'
 import ErrorAlert from '@/components/shared/errorAlert'
 import { calculateStatus } from '@/lib/attendance'
 import { getEventOfTheDay } from '@/actions/event'
