+++ components/admin/usersTable.tsx
@@ -1,6 +1,6 @@
 import { getAllUsers } from '@/queries/select'
 import { DataTable } from '@/components/ui/data-table'
-import { AdminUserColumns } from './adminUserColumns'
+import { AdminUserColumns } from '@/components/admin/adminUserColumns'
 
 export default async function UsersTable() {
   const userRecords = await getAllUsers()
