+++ components/home/onboardingPrompt.tsx
@@ -1,7 +1,7 @@
 import { currentUser } from '@clerk/nextjs/server'
-import OnboardingForm from './onboardingForm'
-import { CardContent } from '../ui/card'
-import ErrorAlert from '../shared/errorAlert'
+import OnboardingForm from '@/components/home/onboardingForm'
+import { CardContent } from '@/components/ui/card'
+import ErrorAlert from '@/components/shared/errorAlert'
 
 export default async function OnboardingPrompt() {
   const user = await currentUser()
