+++ components/shared/errorAlert.tsx
@@ -1,5 +1,5 @@
 import { IconExclamationCircle } from '@tabler/icons-react'
-import { Alert, AlertDescription, AlertTitle } from '../ui/alert'
+import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
 
 export default function ErrorAlert({
   title,
